Start game implementation.

+++ src/main/java/pl/grsrpg/entity/Boss.java
@@ -1,7 +1,7 @@
 package pl.grsrpg.entity;
 
 public class Boss extends Enemy {
-    public Boss(String name, float baseMaxHealth, float baseStrength, float baseAgility, float baseMagicPoints) {
+    public Boss(String name, int baseMaxHealth, int baseStrength, int baseAgility, int baseMagicPoints) {
         super(name, baseMaxHealth, baseStrength, baseAgility, baseMagicPoints);
     }
 }
